feat: convert to typescript
BREAKING CHANGE: convert codebase to typescript
